Add spi nand and flash drivers for realtek SoCs
Renamed the previous patches for rtl9607c platform and clk driver due to
an issue arising from them appearing before rtl838x patches.
Added the luna_mtd_nand for the time being as it contains Realtek specific
drivers for spi nand and flash. Planned to be replaced with something
more mainline linux friendly in the future.

Signed-off-by: Rustam Adilov <adilov@tutamail.com>