realtek: Complete overhaul of the rtl9607c platform
This commit entirely changes the whole rtl9607c platform code. Instead,
It now relies on MACH_REALTEK_RTL config and rtl838x platform code.
The rtl9607c folder, soc drivers and luna_mtd_nand drivers are now deleted.
The clk driver has been refactored and a new patch was addedt to add rtl9607c.
The dtsi file and dts file for the rtl9607c based router are also changed.
The prom in rtl838x modified to add a respective rtl9607c soc details

With all these changes, you can succesfully build the initramfs image which
can be booted into the console on the rtl9607c based router. Though,
there seems to be some issues when during the boot, Illegal instructions
pop up in the bootlog.

The snippet of the respective bootlog:

[    1.063368] Run /init as init process
[    1.067454]   with arguments:
[    1.070753]     /init
[    1.073319]   with environment:
[    1.076817]     HOME=/
[    1.079439]     TERM=linux
Illegal instruction
Illegal instruction
Illegal instruction
[    1.439387] init: Console is alive
[    1.443707] init: - watchdog -
[    1.458449] init: - preinit -

I assume there could be some driver present that is causing that, or
the mtune=4kec might be the culprit. Looks like it is caused by the
patch that has #ifdef CONFIG_RTL838X with hashtag at the start which
doesn't work in arch/mips/Makefile as it treats that as comment so
the inside cflags is active all the time.

Still, a great progress on porting over the rtl9607c platform to OpenWrt.

Signed-off-by: Rustam Adilov <adilov@tutamail.com>